requirements.txt
Django
django-registration-redux
django-colorful